- Changed font file extensions in pubspec.yaml from .ttf to .otf for better compatibility.
- Added new localization keys for 'search' and 'filter' in English, Russian, and Uzbek language files.
- Enhanced HomeScreen layout with a search field and filter button, improving user interaction.
- Refactored CustomCardWidget and CustomBottomSheet to utilize new theme colors and improve UI consistency.
- Introduced a search controller in HomeState for managing search input.

+++ lib/src/feature/user/profile/presentation/widget/bank_card_widget.dart
@@ -12,7 +12,7 @@ class BankCardWidget extends StatelessWidget {
   const BankCardWidget({required this.fullName, required this.balance, required this.id, super.key});
 
   final String fullName;
-  final int balance;
+  final double balance;
   final String id;
 
   @override
@@ -39,7 +39,10 @@ class BankCardWidget extends StatelessWidget {
                 Text('Balance', style: context.textTheme.nunitoW500s14.copyWith(color: context.color.white)),
               ],
             ),
-            Text(balance.toUZSString(), style: context.textTheme.nunitoW600s30.copyWith(color: context.color.white)),
+            Text(
+              balance.toInt().toUZSString(),
+              style: context.textTheme.nunitoW600s30.copyWith(color: context.color.white),
+            ),
             Row(
               mainAxisAlignment: .spaceBetween,
               children: [
